Intensität des Lagerfeuers angepasst an Grösse des Gebäudes Versuch Absturz zu beheben Wachstum von Bäumen verringert DepthOfField entfernt Mainmenu angepasst mit neuen Modellen

+++ Unity-Projekt/Assets/Scripts/GameResources/ResourceData.cs
@@ -17,8 +17,8 @@ public class ResourceData : DatabaseData
 
     // Food
     public bool edible;
-    public int nutrition;
-    public int health;
+    public float nutrition;
+    public float health;
 
     // Processing and crafting
     public float processTime;
